Rename AuthyProvider to AuthyTokenProvider
